requirements.txt
boto3
psycopg2-binary
pandas
streamlit
plotly
python-dotenv
pytz